Determine lower timestamp unit (us|ns) from libpcap magic number
According to [1], the magic number in the libpcap header differs
depending on the timestamp precision:

- 0xA1B2C3D4 indicates records with seconds + microseconds timestamps

- 0xA1B23C4D indicates records with seconds + nanoseconds timestamps

The commit uses this fact to determine the required "unit_modifier" in
"libmoon/lua/pcap.cpp". To do this, the magic_number is directly
passed to the respective functions inside pcap.cpp. This has the
advantage that the pcap header structs in pcap.lua and pcap.cpp don't
need to be modified and continue to conform to the linked Wireshark
source.

[1] https://ietf-opsawg-wg.github.io/draft-ietf-opsawg-pcap/draft-ietf-opsawg-pcap.html